requirements.txt
cachelib                0.7.0
click                   8.1.3
cmake                   3.22.4
colorama                0.4.4
cs50                    9.0.0
dlib                    19.22.99
face-recognition        1.3.0
face-recognition-models 0.3.0
Flask                   2.1.2
Flask-Session           0.4.0
greenlet                1.1.2
itsdangerous            2.1.2
Jinja2                  3.1.2
MarkupSafe              2.1.1
numpy                   1.22.4
opencv-python           4.5.5.64
Pillow                  9.1.1
pip                     22.0.4
setuptools              58.1.0
SQLAlchemy              1.4.36
sqlparse                0.4.2
termcolor               1.1.0
Werkzeug                2.1.2
wheel                   0.37.1